Changed code structure
