+++ src/ml-service/packages/sensor_location/app/modules/locate_model.py
@@ -1,44 +1,7 @@
-import geopandas as gpd
-from shapely.ops import cascaded_union
-from shapely import wkt
-from shapely.geometry import Point
 import pandas as pd
-import numpy as np
-from sklearn.preprocessing import StandardScaler
-from sklearn.model_selection import train_test_split
-from sklearn.cluster import KMeans
-import json
 from pymongo import MongoClient
-
-def process_data(data):
-    #dropping columns with _pc in their names
-    for column in data.columns:
-        if '_pc' in column:
-            data.drop(column, axis = 1, inplace = True)
-    
-    #Renaming some columns for better understandability
-    data.rename(columns = {'d': 'district', 's': 'subcounty', 'p':'parish', 'pop':'population', 'hhs':'households'}, inplace = True)
-    
-    #Drop rows with null values
-    data.dropna(axis=0, inplace=True)
-    data = data.reset_index(drop=True)
-
-    #extracting x, y and z coordinates to replace latitude and longitude 
-    data['coord_x'] = np.cos(data['lat']) * np.cos(data['long'])
-    data['coord_y'] = np.cos(data['lat']) * np.sin(data['long'])
-    data['coord_z'] = np.sin(data['lat'])
-    return data
-
-
-def scaling(data):
-    '''
-    Normalizes data
-    '''
-    data = StandardScaler().fit_transform(data)
-    return data
-
-def split_data(X, y):
-    return train_test_split(X, y, test_size=0.3, shuffle=True)
+from flask import jsonify
+import json
 
 def connect_mongo(db_name):
     '''
@@ -58,7 +21,7 @@ def csv2mongo(csv_path, db_name, collection_name):
     payload=json.loads(data.to_json(orient='records'))
 
     for i in payload:
-        db.kampala.insert_one(i)
+        db[collection_name].insert_one(i)
     
     #db.collection_name.insert_many(payload)
 
@@ -100,3 +63,4 @@ def read_mongo(db_name, collection_name, query={}, no_id=True):
 
 
     
+
